[backend012](temp) Add find free slot function
The function is not completed now and only support
one_time task, the repeated searching is not ready.

Added the apiTest.py about how to use this api,
since our APIs are more and more, I think it's time
to refactor part of our code. (maybe after combining
the code with frontend part)