[debugging] Fix the invite Code API, Add / Update
Before this commit, the inviteCodeAPI only support update
the calendar's members and guests list, which requires user
to firstly obtain the previous list before update.

This commit fixes this bug by adding one extra api:
$ api/scheduler/invitecode/xxx/add/
which allow user to append new member or guest to existing list.

for the old API:
$ api/scheduler/invitecode/xxx/
It will update(overlap) the list, so it could be used for deleting
or managing members and guests list. (Add creating new invite code)